checkpoint: renames "URLComponent/URLPath.ts" to "URLComponent/Path.ts"
